fix(gam-sae): widen matrix-free htbeta slab for flat-chart isometry GN cross-block; fix penalty-FD harness frozen-gate fidelity
Two root-caused fixes plus a #780 file split, all in service of the RED
gam-sae penalty-gradient suite (#1026 reconstruction-parity track).

Production fix — isometry cross-block htbeta width
  `add_sae_isometry_metric_gn_blocks` scatters a Gauss-Newton t↔B
  cross-block into the per-row dense `htbeta` slab and flips on
  `activate_dense_htbeta_supplement`, but ONLY for flat charts (Euclidean,
  Circle, flat products) whose manifold preserves cross-block coherence. On
  the non-frames matrix-free path that slab is allocated at ZERO width
  (#1406/#1407 anti-leak: the data-fit cross-block rides the Kronecker row
  operator, and a `beta_dim = K·M·p` slab was the high-K ~6 TiB leak). With
  an isometry penalty on a flat chart the dense supplement then had nowhere
  to land — the write indexed a `[*,0]` slab and panicked, and dropping it
  would forfeit the strong t↔B coupling the circle fit needs for KKT
  stationarity (#795). New predicate
  `registry_writes_dense_isometry_cross_block` widens the slab to `beta_dim`
  exactly when that dense supplement will be written, and keeps zero width
  otherwise (matrix-free data-fit-only and fixed-decoder paths unchanged).

Test-harness fidelity fix — frozen collinearity gates across FD perturbation
  `assemble_arrow_schur` freezes the #1026 decoder-repulsion and #1625
  separation-barrier collinearity gates at assembly entry and holds them
  fixed across the inner Newton line search; the assembled `gb` is the
  gradient of the frozen-gate objective. In production the line search
  restores only `SaeManifoldMutableState` per trial and leaves the gates
  untouched. The FD probe perturbed via `SaeManifoldTerm::clone`, whose
  custom impl RESETS both frozen gates to `None` — so
  `decoder_repulsion_value` (no live fallback, unlike the barrier) returned
  0 at every perturbed point and the decoder-block FD silently omitted the
  frozen-gate repulsion gradient `gb` carries (~5e-6 on the K=2
  near-collinear fixture). `sae_pen_fd_check` now re-installs the base term's
  frozen gates on every perturbation clone, mirroring the optimizer's
  snapshot/restore. With this, the assembled data-fit+repulsion gradient
  matches central FD to machine precision.

#780 split
  Moved the penalty-FD cluster (`SaeFd*`, `sae_fd_*`, `sae_pen_*`,
  `sae_assembled_gradient_matches_penalized_objective_central_fd`,
  `sae_d1_assembled_gradient_matches_loss_central_fd`,
  `sae_reml_extra_penalty_energy_counts_live_isometry_once`) verbatim into a
  sibling `tests_pen_fd_780.rs` module, dropping tests.rs back under the
  10k-line build-gate ceiling.
